+++ Web_App/Rest/Order/Model/OrderDetailsProductModel.cs
@@ -5,8 +5,8 @@
         public int IdProduct { get; set; }
         public string NameProduct { get; set; }
         public int CountProduct { get; set; }
-        public float ProductCost { get; set; }
-        public float ProductCostSummary { get; set; }
+        public double ProductCost { get; set; }
+        public double ProductCostSummary { get; set; }
         public byte[] ImageUrl { get; set; }
     }
 }
